books_recon/pyproject.toml
[project]
name = "books-recon"
version = "0.1.0"
description = "Evidence-led bookkeeping reconstruction and controlled posting foundation."
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "pydantic-settings>=2.7.0",
  "psycopg[binary]>=3.2.3",
  "python-dotenv>=1.0.1",
  "sqlalchemy>=2.0.36",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
